# Backend/src/api/openapi_3.yaml
...
openapi: 3.0.2
info:
  title: Ausgabenmanager
  version: 1.0.0
  description: "Die Ausgaben-App nutzt einen simplen REST-Webservices zur Kommunikation\
    \ mit dem Backend. Dieser ermöglicht die üblichen CRUD-Operationen zur Suche,\
    \ Anzeige, Bearbeitung und Löschung von Ausgaben, wobei jede Ausgabe aus einem\
    \ Name, Details, Betrag und eine Priorität besteht"
  contact:
    name: "Paul Dittrich, Yannick Elm"
    email: pauldittrich02gmail.com
  license:
    name: The Unlicense
    url: http://unlicense.org
paths:
...
  /income/{incomeId}:
    summary: Path used to manage a single Income.
    description: "The REST endpoint/path used to get, update, and delete single instances\
      \ of an `Income`.  This path contains `GET`, `PUT`, and `DELETE` operations\
      \ used to perform the get, update, and delete tasks, respectively."
    get:
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Income'
          description: Successful response - returns a single `Income`.
      operationId: getIncome
      summary: Get a Income
      description: Gets the details of a single instance of a `Income`.
    put:
      requestBody:
        description: Updated `Income` information.
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Income'
        required: true
      responses:
        "202":
          description: Successful response.
      operationId: updateIncome
      summary: Update a Income
      description: Updates an existing `Income`.
    delete:
      responses:
        "204":
          description: Successful response.
      operationId: deleteIncome
      summary: Delete a Income
      description: Deletes an existing `Income`.
    patch:
      requestBody:
        content:
          application/json: {}
      responses:
        "200":
          description: |
            OK
    parameters:
    - name: incomeId
      description: A unique identifier for a `Income`.
      schema:
        type: string
      in: path
      required: true
components:
  schemas:
    HATEOAS_Link:
      title: Root Type for HATEOAS_Link
      description: HATEOAS-Link
      type: object
      properties:
        url:
          type: string
        method:
          type: string
        query_params:
          type: array
          items:
            type: string
      example:
        url: /expense
        method: GET
        query_params:
        - id
        - name
        - amount
    HATEOAS_Resource:
      title: Root Type for HATEOAS_Resource
      description: HATEOAS-Links einer Resource
      type: object
      properties:
        read:
          $ref: '#/components/schemas/HATEOAS_Link'
        update:
          $ref: '#/components/schemas/HATEOAS_Link'
        patch:
          $ref: '#/components/schemas/HATEOAS_Link'
        delete:
          $ref: '#/components/schemas/HATEOAS_Link'
      example:
        read: ""
        update: ""
        patch: ""
        delete: ""
...
    Income:
      description: Eintrag in Ausgaben-Manager
      type: object
      properties:
        _links:
          $ref: '#/components/schemas/HATEOAS_Resource'
          description: ""
...